move EpochChangesObserver to data
