+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/data/conversation/EpochChangesObserver.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
-package com.wire.kalium.logic.feature.conversation.mls
+package com.wire.kalium.logic.data.conversation
 
 import com.wire.kalium.logic.data.id.GroupID
 import kotlinx.coroutines.flow.Flow
